Fixed initial issues

+++ deploy/50-cleanup/10-clear-old-builds.sh
@@ -13,19 +13,37 @@ if [ -f "${BUILD_ID}.tar.gz" ]; then
 fi
 
 NUM=0
-PREVIOUS_BUILDS=2
+PREVIOUS_BUILDS=1
+shopt -s nullglob
 for FILE in `ls -d -- ${RELEASE_PARENT_DIR}/*-* | sort -t- -k2 -r`
 do
+    if [[ ! -d ${FILE} ]]; then
+        continue
+    fi
     if [[ ${FILE} != *"-${BUILD}"* ]]; then
+        FILE_BASE=`basename ${FILE}`
+        IFS='-' read -ra BUILD_ARR <<< "$FILE_BASE"
+        BUILD_LABEL=${BUILD_ARR[0]}
+        BUILD_NUM=${BUILD_ARR[1]}
+        re='^[0-9]+$'
+        if ! [[ $BUILD_NUM =~ $re ]]; then
+            printf "${FILE_BASE} is not a build directory\n"
+            continue
+        fi
+        if ! [[ "$BUILD_LABEL" = "build" ]]; then
+            printf "${FILE_BASE} is not a build directory\n"
+            continue
+        fi
         let "NUM=NUM+1"
         if [[ $NUM > $PREVIOUS_BUILDS ]]; then
             printf "DELETING ${FILE}\n"
             rm -rf ${FILE}
         else
-            IFS='-' read -ra BUILD_ARR <<< "$FILE"
-            printf "Keeping recent build ${BUILD_ARR[1]}\n"
+            printf "Keeping recent build ${BUILD_NUM}\n"
         fi
     else
         printf "Keeping current build ${BUILD}\n"
     fi
 done
+
+rm -rf ${RELEASE_PARENT_DIR}/*.gz
